削除機能の修正

+++ app/controllers/protospaces_controller.rb
@@ -39,8 +39,12 @@ class ProtospacesController < ApplicationController
 
     def destroy
         prototype = Prototype.find(params[:id])
-        prototype.destroy
-        redirect_to root_path        
+        if current_user != prototype.user
+            redirect_to root_path
+        else
+            prototype.destroy
+            redirect_to root_path                 
+        end   
     end
 
     private
